convert euifacetbutton to ts
